closes #355 - remove adminLesson data type, refactor code to use lesson type from graphql/index.tsx file
