--- @types/adminLesson.d.ts
@@ -1,10 +1,0 @@
-export type Lesson = {
-  id: string
-  title: string
-  description: string
-  order: number
-  challenges: Challenge[]
-  docUrl: string
-  chatUrl: string
-  __typename: string
-}
